backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pandas==2.2.2
scikit-learn==1.4.2
networkx==3.3
python-dotenv==1.0.1
python-multipart==0.0.9
google-generativeai==0.8.4
pydantic==2.7.1
numpy==1.26.4
